==> notebooks/00-split_dataset.ipynb <==
This notebook is intended to create a small dataset to simulate new customers and test all solution build in this project.

import pandas as pd

# Original dataset from Kaggle
df = pd.read_csv("../data/hotel_bookings.csv")

# Dataset shape
df.shape

# Count the number of records by hotel
df.groupby("hotel")['hotel'].count()

# Randomly select 200 records
df_newcustomers = df.sample(n=10000, random_state=1)  # random_state for reproducibility

# Create the dataset with the remaining 9800 records
df_dataset = df.drop(df_newcustomers.index)

df_newcustomers.groupby("hotel")['hotel'].count(), df_dataset.groupby("hotel")['hotel'].count()

# Save the new datasets to CSV files
df_newcustomers.to_csv('../data/new_customers.csv', index=False)
df_dataset.to_csv('../data/hotel_booking_dataset.csv', index=False)

# Print a summary of the created datasets
print(f'Original dataset: {df.shape[0]} records')
print(f'Sample dataset: {df_newcustomers.shape[0]} records')
print(f'Remaining dataset: {df_dataset.shape[0]} records')
