==> notebooks/03-test-inferences.ipynb <==
from mlflow.tracking import MlflowClient
import mlflow.sklearn
import pandas as pd

# Configura la URL del servidor de MLflow
mlflow.set_tracking_uri("http://localhost:5000")

# Crea un cliente de MLflow
client = MlflowClient()

# Lista todos los modelos registrados
models = client.search_registered_models()

# Imprime los nombres de los modelos registrados
print("Modelos registrados en el servidor de MLflow:")
for model in models:
    print(f"Nombre del modelo: {model.name}")

model_name = "sk-learn-logistic-regression-reg-model"
#model_alias = "champion"
model_alias = "challenger"

model_version_details = client.get_model_version_by_alias(model_name, model_alias)

artifact_path = 'preprocessing/preprocessor_model.pkl'
local_path = mlflow.artifacts.download_artifacts(run_id=model_version_details.run_id, artifact_path=artifact_path, dst_path='./assets')

# Carga el modelo usando el alias
model_uri = f"models:/{model_name}@{model_alias}"
model = mlflow.sklearn.load_model(model_uri)

model

model_info = mlflow.models.Model.load(f"runs:/{model_version_details.run_id}/model")
signature = model_info.signature

data = {
    "cat__hotel_City Hotel": [1.0],
    "cat__hotel_Resort Hotel": [0.0],
    "cat__meal_BB": [1.0],
    "cat__meal_FB": [0.0],
    "cat__meal_HB": [0.0],
    "cat__meal_SC": [0.0],
    "cat__meal_Undefined": [0.0],
    "cat__market_segment_Aviation": [0.0],
    "cat__market_segment_Complementary": [0.0],
    "cat__market_segment_Corporate": [0.0],
    "cat__market_segment_Direct": [0.0],
    "cat__market_segment_Groups": [0.0],
    "cat__market_segment_Offline TA/TO": [0.0],
    "cat__market_segment_Online TA": [1.0],
    "cat__market_segment_Undefined": [0.0],
    "cat__distribution_channel_Corporate": [0.0],
    "cat__distribution_channel_Direct": [0.0],
    "cat__distribution_channel_GDS": [0.0],
    "cat__distribution_channel_TA/TO": [1.0],
    "cat__distribution_channel_Undefined": [0.0],
    "cat__reserved_room_type_A": [1.0],
    "cat__reserved_room_type_B": [0.0],
    "cat__reserved_room_type_C": [0.0],
    "cat__reserved_room_type_D": [0.0],
    "cat__reserved_room_type_E": [0.0],
    "cat__reserved_room_type_F": [0.0],
    "cat__reserved_room_type_G": [0.0],
    "cat__reserved_room_type_H": [0.0],
    "cat__reserved_room_type_L": [0.0],
    "cat__reserved_room_type_P": [0.0],
    "cat__deposit_type_No Deposit": [1.0],
    "cat__deposit_type_Non Refund": [0.0],
    "cat__deposit_type_Refundable": [0.0],
    "cat__customer_type_Contract": [0.0],
    "cat__customer_type_Group": [0.0],
    "cat__customer_type_Transient": [1.0],
    "cat__customer_type_Transient-Party": [0.0],
    "num__lead_time": [-0.7091565077403539],
    "num__days_in_waiting_list": [-0.07563313143189111],
    "num__adr": [-0.27708191730285603],
    "num__total_stay": [-0.5903231218339896],
    "num__total_people": [-0.02995204191832186],
    "bin__is_repeated_guest": [0],
    "bin__previous_cancellations": [0],
    "bin__previous_bookings_not_canceled": [0],
    "bin__booking_changes": [0],
    "bin__agent": [1],
    "bin__company": [0],
    "bin__required_car_parking_spaces": [0],
    "bin__total_of_special_requests": [1]
}

df = pd.DataFrame(data)

pred = model.predict(df)
pred_prob = model.predict_proba(df)

pred, pred_prob
