==> notebooks/04-test-inferences-batch.ipynb <==
from mlflow.tracking import MlflowClient
import mlflow.sklearn
import pandas as pd
from datetime import datetime, timedelta
from sqlalchemy import create_engine
from dotenv import load_dotenv
import os
import joblib

load_dotenv(dotenv_path='../.env')

PG_PREDICTION_PWD = os.getenv('PG_PREDICTION_PWD')
url_conexion = f'postgresql+psycopg2://prediction_user:{PG_PREDICTION_PWD}@localhost:5432/reservations_db'

engine = create_engine(url_conexion)

nombre_tabla = 'reservations'
df = pd.read_sql_query(f'SELECT * FROM {nombre_tabla}', engine)

mlflow.set_tracking_uri("http://localhost:5000")

client = MlflowClient()

models = client.search_registered_models()

print("Models registered in MLflow:")
for model in models:
    print(f"Model name: {model.name}")

model_name = "sk-learn-logistic-regression-reg-model"
model_alias = "champion"
model_version_details = client.get_model_version_by_alias(model_name, model_alias)

artifact_path = 'preprocessing/preprocessor_model.pkl'
local_path = mlflow.artifacts.download_artifacts(run_id=model_version_details.run_id, artifact_path=artifact_path, dst_path='./assets')
preprocessor = joblib.load('artifacts/preprocessor_model.pkl')

# Carga el modelo usando el alias
model_uri = f"models:/{model_name}@{model_alias}"
model = mlflow.sklearn.load_model(model_uri)

model

model_info = mlflow.models.Model.load(f"runs:/{model_version_details.run_id}/model")
signature = model_info.signature

# Defining categorical, numerical and binary columns
cat_cols = ['hotel','meal', 'market_segment','distribution_channel',
            'reserved_room_type','deposit_type','customer_type']

num_cols = ['lead_time','days_in_waiting_list',
            'adr','total_stay','total_people']

bin_cols = ['is_repeated_guest','previous_cancellations',
            'previous_bookings_not_canceled','booking_changes',
            'agent','company','required_car_parking_spaces',
            'total_of_special_requests']

dataset =(df
 .drop_duplicates()
 .fillna(0)
 .assign(total_stay=lambda df: df['stays_in_weekend_nights'] + df['stays_in_week_nights'],
         total_people=lambda df: df['adults'] + df['children'] + df['babies'],
        )
 [cat_cols + num_cols + bin_cols]
 .assign(total_people=lambda df: df['total_people'].astype('int64'),
         agent=lambda df: df['agent'].astype('int64'),
         company=lambda df: df['company'].astype('int64'),
        )
)

dataset

new_dataset = preprocessor.transform(dataset)

print(new_dataset)

pred = model.predict(new_dataset)
pred_prob = model.predict_proba(new_dataset)

pred, pred_prob
